+++ my-app/src/App.css
@@ -47,4 +47,8 @@
   justify-content: space-around;
 }
 
+.checkbox {
+  height: 1.5rem;
+  width: 1.5rem;
+}
